use inference util
