+++ python/nano/src/bigdl/nano/tf/keras/__init__.py
@@ -15,10 +15,6 @@
 #
 import tensorflow as tf
 
-try:
-    from ..quantization import quantize
-    tf.keras.Model.quantize = quantize
-except ImportError:
-    print("Warning: Intel Neural Compressor should be installed if you need quantization.")
+
 from .Sequential import Sequential
 from .Model import Model
